+++ lib/decoder.dart
@@ -7,5 +7,4 @@ library toml.decoder;
 export 'src/decoder/exception/invalid_escape_sequence.dart';
 export 'src/decoder/exception/not_a_table.dart';
 export 'src/decoder/exception/redefinition.dart';
-export 'src/decoder/interface.dart';
 export 'src/decoder/map_builder.dart';