+++ lib/toml.dart
@@ -8,4 +8,3 @@ export 'ast.dart';
 export 'decoder.dart';
 export 'encoder.dart';
 export 'exception.dart';
-export 'loader.dart';